requirements.txt
Flask
python-dotenv
google-genai
gunicorn
flask-cors
tavily-python
pypdf
supabase